RUMM-3353 Crash Reporter Configuration
