+++ Tests/DatadogTests/Datadog/RUM/RUMMonitorTests.swift
@@ -1079,7 +1079,7 @@ class RUMMonitorTests: XCTestCase {
         )
 
         // Given
-        DatadogCrashReporter.initialize(in: core)
+        CrashReporter.enable(in: core)
         let crashReporter = try XCTUnwrap(core.get(feature: CrashReportingFeature.self))
 
         try core.register(feature: crashReporter)
