+++ dependency-manager-tests/cocoapods/App/ViewController.swift
@@ -32,7 +32,7 @@ internal class ViewController: UIViewController {
 
         Logs.enable()
 
-        DatadogCrashReporter.initialize()
+        CrashReporter.enable()
 
         self.logger = DatadogLogger.builder
             .sendLogsToDatadog(false)
