+++ dependency-manager-tests/spm/App/ViewController.swift
@@ -30,7 +30,7 @@ internal class ViewController: UIViewController {
 
         Logs.enable()
 
-        DatadogCrashReporter.initialize()
+        CrashReporter.enable()
 
         self.logger = DatadogLogger.builder
             .sendLogsToDatadog(false)
